Extract main.rs's mod real into src/runtime.rs; thin main.rs to 53 lines
main.rs now holds only module declarations, the non-linux stub, and main().
The pre-stream/streaming loop and its helpers move verbatim into runtime (renamed
from real). No behavior change (cargo fmt reindented the moved block).
