+++ src/session/mod.rs
@@ -22,9 +22,9 @@ use punktfunk_core::packet::{FLAG_SOF, USER_FLAG_RECOVERY_ANCHOR};
 use punktfunk_core::quic;
 
 use crate::platform::webos::ndl::{NdlCodec, NdlVideo};
-use crate::session::pacing::{HostPtsAnchor, PtsPacer};
-use crate::services::store::{CodecPref, ColorRangeOverride, VideoBackend};
 use crate::platform::webos::starfish::StarfishVideo;
+use crate::services::store::{CodecPref, ColorRangeOverride, VideoBackend};
+use crate::session::pacing::{HostPtsAnchor, PtsPacer};
 
 impl ColorRangeOverride {
     /// Force the VUI `full_range` flag per the user override before it's handed to
